+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/generation/variables/ExternalArray.java
@@ -27,6 +27,11 @@ public class ExternalArray extends AbstractArrayStore {
     }
 
     @Override
+    public boolean valid() {
+        return true;
+    }
+
+    @Override
     public boolean isDeclaredRemote() {
         return false;
     }