+++ CHANGELOG.markdown
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 * Added tail call optimization to [Recursive Optimization](/doc/syntax/optimizations/RECURSIVE-OPTIMIZATION.markdown#tail-call-optimization).
 * Added an [internal stack implementation](doc/syntax/SYNTAX-4-FUNCTIONS.markdown#internal-stack), which supports storing non-numerical values in parameters and variables of recursive functions without restrictions.
 * Added stack overflow runtime check for both internal and external stack, govedned by the [`stack-overflow-checks` compiler option](/doc/syntax/SYNTAX-5-OTHER.markdown#option-stack-overflow-checks).
+* Added support for declaring arrays in functions (as local variables).
 * Added support for compile-time evaluation of the `min` and `max` varargs functions, and the `ascii`, `char` and `strlen` built-in functions. These functions now can be used in constant expressions.
 
 ### Changed